This is the ROM Hack!
Make sure you patch this with an official Super Mario Bros ROM

VVV BEFORE YOU PLAY! VVV

Some times going to the Factory entrance can bug out and not work properly! I suggest you fully load the factory and then enter so it will not take you to world 5!
This was made for the USA region, so I do not know if it works for the other regions or not!